refactor(ci-hygiene): split badge and hook commands (#0000)
