+++ crates/perl-ci-hygiene/src/main.rs
@@ -10,19 +10,21 @@ use std::collections::{BTreeMap, HashMap, HashSet};
 use std::env;
 use std::ffi::OsStr;
 use std::fs;
-#[cfg(unix)]
-use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::sync::LazyLock;
 use toml::Value as TomlValue;
 use walkdir::{DirEntry, WalkDir};
 
+mod badges;
 mod cli;
 mod commands;
+mod git_hooks;
 mod process;
 
+use crate::badges::cmd_generate_badges;
 use crate::cli::{Cli, CliCommand};
+use crate::git_hooks::cmd_install_githooks;
+#[cfg(test)]
+use crate::git_hooks::pre_push_hook_script;
 use crate::process::*;
 
 const RED: &str = "\x1b[0;31m";
@@ -30,21 +32,6 @@ const GREEN: &str = "\x1b[0;32m";
 const YELLOW: &str = "\x1b[0;33m";
 const BLUE: &str = "\x1b[0;34m";
 const NC: &str = "\x1b[0m";
-
-static VS_MARKETPLACE_INSTALLS_BADGE_RE: LazyLock<Regex> = LazyLock::new(|| {
-    Regex::new(r"VS%20Marketplace-(\d+)%20installs").unwrap_or_else(|error| {
-        unreachable!("VS_MARKETPLACE_INSTALLS_BADGE_RE is a known-good static pattern: {error}")
-    })
-});
-
-static VS_MARKETPLACE_INSTALLS_BADGE_BLOCK_RE: LazyLock<Regex> = LazyLock::new(|| {
-    Regex::new(
-        r"(?s)<!-- perl-lsp:vs-marketplace-installs-badge:start -->.*?<!-- perl-lsp:vs-marketplace-installs-badge:end -->",
-    )
-    .unwrap_or_else(|error| {
-        unreachable!("VS_MARKETPLACE_INSTALLS_BADGE_BLOCK_RE is a known-good static pattern: {error}")
-    })
-});
 
 fn main() -> std::process::ExitCode {
     if let Err(err) = color_eyre::install() {
@@ -1501,457 +1488,6 @@ fn find_cancel_test_binary(repo_root: &Path) -> Option<PathBuf> {
         }
     }
     None
-}
-
-fn cmd_generate_badges(repo_root: &Path, check_mode: bool) -> Result<i32> {
-    let facts_file = repo_root.join("docs/project/publication-facts.toml");
-    let facts_content = fs::read_to_string(&facts_file)
-        .with_context(|| format!("reading facts file {:?}", facts_file))?;
-    let facts: TomlValue = toml::from_str(&facts_content)
-        .with_context(|| format!("parsing facts file {:?}", facts_file))?;
-
-    let vscode_installs_i64 = facts
-        .get("external")
-        .and_then(|e| e.get("vscode_marketplace_installs"))
-        .and_then(|v| v.get("value"))
-        .and_then(|v| v.as_integer())
-        .ok_or_else(|| {
-            color_eyre::eyre::eyre!(
-                "Could not read vscode_marketplace_installs.value from {}",
-                facts_file.display()
-            )
-        })?;
-
-    let vscode_installs = u32::try_from(vscode_installs_i64).with_context(|| {
-        format!("vscode_marketplace_installs value {} is out of range for u32", vscode_installs_i64)
-    })?;
-
-    let badge_url = format!(
-        "https://img.shields.io/badge/VS%20Marketplace-{}%20installs-0078D4",
-        vscode_installs
-    );
-
-    let root_readme = repo_root.join("README.md");
-    let ext_readme = repo_root.join("vscode-extension/README.md");
-
-    if check_mode {
-        let mut has_drift = false;
-        for readme_path in [&root_readme, &ext_readme] {
-            if readme_path.exists() {
-                let content = fs::read_to_string(readme_path)?;
-                if !content.contains(&badge_url) {
-                    eprintln!(
-                        "{}VS Marketplace badge drift in {}{}",
-                        RED,
-                        readme_path.display(),
-                        NC
-                    );
-                    eprintln!(
-                        "  expected installs: {} from {}",
-                        vscode_installs,
-                        facts_file.display()
-                    );
-
-                    if let Some(caps) = VS_MARKETPLACE_INSTALLS_BADGE_RE.captures(&content)
-                        && let Some(found) = caps.get(1)
-                    {
-                        eprintln!(
-                            "  stale badge found: {} but expected {} in {}",
-                            found.as_str(),
-                            vscode_installs,
-                            readme_path.display()
-                        );
-                    }
-                    has_drift = true;
-                }
-            }
-        }
-        if has_drift {
-            eprintln!("Run: cargo xtask ci-hygiene generate-badges");
-            return Ok(1);
-        }
-        println!("{}✓ VS Marketplace badge check passed{}", GREEN, NC);
-        return Ok(0);
-    }
-
-    // Generate mode: update badges
-    for readme_path in [&root_readme, &ext_readme] {
-        if !readme_path.exists() {
-            continue;
-        }
-
-        let content = fs::read_to_string(readme_path)?;
-        let updated = update_badge_in_content(&content, &badge_url)?;
-
-        if updated != content {
-            fs::write(readme_path, &updated)
-                .with_context(|| format!("writing updated badge to {:?}", readme_path))?;
-            println!("{}✓ Updated VS Marketplace badge in {}{}", GREEN, readme_path.display(), NC);
-        }
-    }
-
-    println!(
-        "{}✓ Badges updated from value {} in publication-facts.toml{}",
-        GREEN, vscode_installs, NC
-    );
-    Ok(0)
-}
-
-fn update_badge_in_content(content: &str, badge_url: &str) -> Result<String> {
-    if !VS_MARKETPLACE_INSTALLS_BADGE_BLOCK_RE.is_match(content) {
-        return Ok(content.to_string());
-    }
-
-    // Determine if we're dealing with HTML or Markdown format
-    if content.contains("href=\"https://marketplace.visualstudio.com") {
-        // HTML format
-        let replacement = format!(
-            "<!-- perl-lsp:vs-marketplace-installs-badge:start -->\n  <a href=\"https://marketplace.visualstudio.com/items?itemName=EffortlessMetrics.perl-lsp-rs\"><img src=\"{}\" alt=\"VS Marketplace installs\" /></a>\n  <!-- perl-lsp:vs-marketplace-installs-badge:end -->",
-            badge_url
-        );
-        return Ok(VS_MARKETPLACE_INSTALLS_BADGE_BLOCK_RE
-            .replace_all(content, replacement)
-            .into_owned());
-    }
-
-    // Markdown format
-    let replacement = format!(
-        "<!-- perl-lsp:vs-marketplace-installs-badge:start -->\n[![VS Marketplace Installs (manual)]({})](https://marketplace.visualstudio.com/items?itemName=EffortlessMetrics.perl-lsp-rs)\n<!-- perl-lsp:vs-marketplace-installs-badge:end -->",
-        badge_url
-    );
-    Ok(VS_MARKETPLACE_INSTALLS_BADGE_BLOCK_RE.replace_all(content, replacement).into_owned())
-}
-
-fn pre_push_hook_script() -> &'static str {
-    r#"#!/usr/bin/env bash
-# ============================================================================
-# perl-lsp pre-push hook (generated by `cargo xtask ci-hygiene install-githooks`)
-# ============================================================================
-#
-# Bypass policy
-# -------------
-# OK to bypass with `git push --no-verify`:
-#   * Deletion-only push on master (the hook already auto-skips, but if it
-#     doesn't, bypassing is safe — there's nothing to validate).
-#   * Urgent fixes during incident response, when the gate has a known bug
-#     being tracked (see hint output below for issue numbers).
-#   * The hook is failing for an environmental reason that is out of band of
-#     your change (e.g., nix store cache miss, transient toolchain issue).
-#
-# NOT OK to bypass:
-#   * "I just don't want to wait."
-#   * "I just want to push something quick."
-#   * Code-touching changes where you haven't actually run the gate locally.
-#
-# If you find yourself bypassing repeatedly, file an issue and link it here.
-# ============================================================================
-
-set -euo pipefail
-
-# --- Self-heal core.bare corruption (issue #3205) ---
-# Some sequences of `git worktree add`/`remove` silently flip core.bare=true
-# on the main checkout, breaking every worktree-aware git command including
-# this hook. Auto-unset before doing anything else.
-if [ "$(git config --get core.bare 2>/dev/null || true)" = "true" ]; then
-    # Confirm this is actually a non-bare repo (has a working tree).
-    if git rev-parse --show-toplevel >/dev/null 2>&1; then
-        echo "⚠️  Detected core.bare=true corruption (issue #3205) — auto-fixing"
-        git config --local --unset core.bare || true
-    fi
-fi
-
-# --- Self-heal stale hook installation (issue #4220) ---
-# When hooks/pre-push is updated in master, .git/hooks/pre-push is only
-# updated when install-githooks is re-run. Auto-copy when drift is detected.
-# Note: exec "$0" "$@" does NOT work here — git stdin is already consumed
-# before the hook executes. Copy and continue; the fresh hook takes effect
-# on the next push.
-REPO_ROOT_FOR_HOOK="$(git rev-parse --show-toplevel 2>/dev/null || true)"
-if [ -n "$REPO_ROOT_FOR_HOOK" ] && [ -f "$REPO_ROOT_FOR_HOOK/hooks/pre-push" ]; then
-    if ! diff -q "$0" "$REPO_ROOT_FOR_HOOK/hooks/pre-push" >/dev/null 2>&1; then
-        echo "pre-push hook updated from hooks/pre-push (was stale — takes effect next push)"
-        cp "$REPO_ROOT_FOR_HOOK/hooks/pre-push" "$0" && chmod +x "$0" || true
-    fi
-fi
-
-# stdin provides: <local ref> <local sha> <remote ref> <remote sha>
-# Git sends all-zero SHA for deletions. Read all refs into an array first
-# so stdin is available for both the delete-check and the test-file scan.
-PUSH_REFS=()
-while IFS= read -r line; do
-    PUSH_REFS+=("$line")
-done
-
-# --- Skip CI gate when all refs are being deleted ---
-IS_DELETE_ONLY=true
-for line in "${PUSH_REFS[@]+"${PUSH_REFS[@]}"}"; do
-    local_sha="$(echo "$line" | awk '{print $2}')"
-    if [ "$local_sha" != "0000000000000000000000000000000000000000" ]; then
-        IS_DELETE_ONLY=false
-        break
-    fi
-done
-
-if [ "$IS_DELETE_ONLY" = true ]; then
-    echo "Branch deletion — skipping CI gate"
-    exit 0
-fi
-
-# --- Detect doc-only changes for the fast-path gate ---
-# A push is doc-only if every changed file matches one of:
-#   *.md, *.txt, LICENSE*, CHANGELOG*, docs/**, .github/ISSUE_TEMPLATE/**,
-#   or */LICENSE* (crate-subdir license files, e.g. crates/*/LICENSE-APACHE)
-# Doc-only pushes skip code gates entirely instead of running the full
-# ci-gate, since the test suite and workspace-wide rustfmt check are
-# pointless for prose-only changes and can fail spuriously on Windows (#4047).
-REPO_ROOT="$(git rev-parse --show-toplevel)"
-DOC_ONLY=true
-TEST_FILES_CHANGED=false
-HAS_DIFFABLE_REF=false
-# Track unique crate names for the single-crate tier.
-# We use a newline-separated string rather than an array for POSIX compat.
-SINGLE_CRATE_NAMES=""
-SINGLE_CRATE_ALL_UNDER_CRATES=true
-for line in "${PUSH_REFS[@]+"${PUSH_REFS[@]}"}"; do
-    read -r _local_ref local_sha _remote_ref remote_sha <<< "$line"
-    # For new branches, compare against merge-base with origin/master
-    if [ "$remote_sha" = "0000000000000000000000000000000000000000" ]; then
-        remote_sha="$(git merge-base "$local_sha" origin/master 2>/dev/null || echo "$local_sha")"
-    fi
-    CHANGED_FILES="$(git diff --name-only "$remote_sha" "$local_sha" 2>/dev/null || true)"
-    if [ -n "$CHANGED_FILES" ]; then
-        HAS_DIFFABLE_REF=true
-    fi
-    while IFS= read -r changed; do
-        [ -z "$changed" ] && continue
-        case "$changed" in
-            *.md|*.txt|LICENSE*|CHANGELOG*|docs/*|.github/ISSUE_TEMPLATE/*|*/LICENSE*)
-                ;;
-            *)
-                DOC_ONLY=false
-                # Track whether this code file lives under crates/<name>/
-                case "$changed" in
-                    crates/*/*)
-                        # Extract the crate directory name: crates/<name>/...
-                        crate_name="${changed#crates/}"
-                        crate_name="${crate_name%%/*}"
-                        # Append to list if not already present
-                        if ! printf '%s\n' "$SINGLE_CRATE_NAMES" | grep -qxF "$crate_name" 2>/dev/null; then
-                            SINGLE_CRATE_NAMES="${SINGLE_CRATE_NAMES}${crate_name}
-"
-                        fi
-                        ;;
-                    *)
-                        # File is outside crates/ — can't be single-crate
-                        SINGLE_CRATE_ALL_UNDER_CRATES=false
-                        ;;
-                esac
-                ;;
-        esac
-    done <<< "$CHANGED_FILES"
-    if echo "$CHANGED_FILES" | grep -qE '^crates/.*/tests/.*\.rs$'; then
-        TEST_FILES_CHANGED=true
-    fi
-done
-
-# If we couldn't compute a diff for any ref (e.g., shallow clone, missing
-# remote), assume code changes and run the full gate to be safe.
-if [ "$HAS_DIFFABLE_REF" != true ]; then
-    DOC_ONLY=false
-fi
-
-if [ "$DOC_ONLY" = true ]; then
-    echo "📝 Doc-only push — skipping code gates"
-    echo "   (Skip with: git push --no-verify)"
-    echo "✅ Doc-only fast-path gate passed"
-    exit 0
-fi
-
-# --- Single-crate proportional gate ---
-# If every code change is under a single crates/<name>/ directory, run a
-# targeted cargo fmt/clippy/test -p <name> instead of the full workspace gate.
-# Falls back to the full gate if classification is ambiguous.
-SINGLE_CRATE_COUNT="$(printf '%s' "$SINGLE_CRATE_NAMES" | grep -c . 2>/dev/null || echo 0)"
-if [ "$SINGLE_CRATE_ALL_UNDER_CRATES" = true ] && [ "$SINGLE_CRATE_COUNT" = "1" ]; then
-    SINGLE_CRATE_DIR="$(printf '%s' "$SINGLE_CRATE_NAMES" | tr -d '[:space:]')"
-    # Resolve the Cargo package name from Cargo.toml, not the directory basename.
-    # This fixes the perl-lsp -> perl-lsp-rs mismatch (issue #4512).
-    # Falls back to directory name if cargo xtask is unavailable.
-    SINGLE_CRATE_NAME="$(cargo xtask resolve-package-name "crates/${SINGLE_CRATE_DIR}" 2>/dev/null || printf '%s' "$SINGLE_CRATE_DIR")"
-    echo "Single-crate push (${SINGLE_CRATE_DIR} -> ${SINGLE_CRATE_NAME}) — running targeted gate"
-    echo "   (Skip with: git push --no-verify)"
-    echo ""
-    run_single_crate_gate() {
-        cargo fmt -p "$SINGLE_CRATE_NAME" -- --check && \
-        cargo clippy -p "$SINGLE_CRATE_NAME" -- -D warnings && \
-        cargo test -p "$SINGLE_CRATE_NAME"
-    }
-    GATE_LOG="$(mktemp -t perl-lsp-prepush.XXXXXX.log 2>/dev/null || mktemp)"
-    trap 'rm -f "$GATE_LOG"' EXIT
-    set +e
-    run_single_crate_gate 2>&1 | tee "$GATE_LOG"
-    GATE_STATUS=${PIPESTATUS[0]}
-    set -e
-    if [ "$GATE_STATUS" -ne 0 ]; then
-        echo ""
-        echo "❌ Single-crate gate failed (exit $GATE_STATUS)"
-        echo "   See bypass policy at the top of .git/hooks/pre-push for when"
-        echo "   --no-verify is appropriate."
-        exit "$GATE_STATUS"
-    fi
-    echo "✅ Single-crate gate passed"
-    exit 0
-fi
-
-echo "Running local fast gate before push: nix develop -c just pr-fast"
-echo "   (Skip with: git push --no-verify)"
-echo ""
-
-# --- Check if test files changed and CURRENT_STATUS.md needs updating ---
-if [ "$TEST_FILES_CHANGED" = true ] && [ -f "$REPO_ROOT/scripts/update-current-status.py" ]; then
-    echo "📊 Test files changed — checking if docs/project/status/ is up to date..."
-    if command -v python3 &>/dev/null; then
-        python3 "$REPO_ROOT/scripts/update-current-status.py" 2>/dev/null || true
-        if ! git diff --quiet -- docs/project/status/ 2>/dev/null; then
-            echo ""
-            echo "⚠️  docs/project/status/ has stale test counts!"
-            echo "   Run: python3 scripts/update-current-status.py"
-            echo "   Then commit the updated files before pushing."
-            echo "   git add docs/project/status/"
-            echo ""
-            exit 1
-        fi
-    fi
-fi
-
-# --- Run the full gate, capturing output for failure-mode hints ---
-GATE_LOG="$(mktemp -t perl-lsp-prepush.XXXXXX.log 2>/dev/null || mktemp)"
-trap 'rm -f "$GATE_LOG"' EXIT
-
-run_gate() {
-    if command -v nix &>/dev/null && [ -f flake.nix ]; then
-        nix develop -c just pr-fast
-    elif command -v just &>/dev/null; then
-        just pr-fast
-    else
-        echo "⚠️  Neither 'nix develop' nor 'just' available, skipping pre-push gate"
-        echo "   Install just: cargo install just"
-        return 0
-    fi
-}
-
-set +e
-run_gate 2>&1 | tee "$GATE_LOG"
-GATE_STATUS=${PIPESTATUS[0]}
-set -e
-
-if [ "$GATE_STATUS" -ne 0 ]; then
-    echo ""
-    echo "❌ pr-fast failed (exit $GATE_STATUS) — checking for known issues..."
-    HINTED=false
-    if grep -q 'ci-parser-features-check' "$GATE_LOG" 2>/dev/null && \
-       grep -qE 'os error 5|Access is denied' "$GATE_LOG" 2>/dev/null; then
-        echo "   • Known: Windows file-lock race in ci-parser-features-check (#3202)"
-        echo "     Workaround: re-run the gate, or use --no-verify if you're confident"
-        echo "     your change is unrelated to xtask/parser-features."
-        HINTED=true
-    fi
-    if grep -qE 'cargo (xtask )?fmt.*--check' "$GATE_LOG" 2>/dev/null && \
-       grep -qE 'Diff in|rustfmt' "$GATE_LOG" 2>/dev/null; then
-        echo "   • Formatting drift — run \`cargo xtask fmt\` to auto-fix"
-        HINTED=true
-    fi
-    if grep -qE 'clippy::|warning: .*-> .*\.rs' "$GATE_LOG" 2>/dev/null; then
-        echo "   • Clippy warnings — look for the error above and fix the warnings"
-        HINTED=true
-    fi
-    if grep -qE 'os error 206|filename.*too long|ERROR_FILENAME_EXCED' "$GATE_LOG" 2>/dev/null; then
-        echo "   • Windows CreateProcess command-line length limit (os error 206)"
-        echo "     'cargo fmt --all' passes all 1200+ source files in one command,"
-        echo "     exceeding the ~32K-char CreateProcess limit on Windows."
-        echo "     Fix: bash scripts/install-githooks.sh"
-        echo "     This installs the current hook which uses 'cargo xtask fmt' (per-crate)."
-        echo "     See: docs/contributing/FIRST_PR.md"
-        HINTED=true
-    fi
-    if [ "$HINTED" = false ]; then
-        echo "   No known-issue patterns matched. Read the gate output above."
-    fi
-    echo ""
-    echo "   See bypass policy at the top of .git/hooks/pre-push for when"
-    echo "   --no-verify is appropriate."
-    exit "$GATE_STATUS"
-fi
-"#
-}
-
-fn cmd_install_githooks(repo_root: &Path) -> Result<i32> {
-    let hooks_dir = resolve_git_hooks_dir(repo_root)?;
-    fs::create_dir_all(&hooks_dir)?;
-
-    let pre_commit_hook = r#"#!/usr/bin/env bash
-set -euo pipefail
-
-GIT_USER_NAME="$(git config user.name 2>/dev/null || true)"
-GIT_USER_EMAIL="$(git config user.email 2>/dev/null || true)"
-
-if [ "$GIT_USER_NAME" = "Codex Release Validation" ] || \
-   [ "$GIT_USER_EMAIL" = "codex-release-validation@example.invalid" ] || \
-   [ "$GIT_USER_NAME" = "xtask hook tests" ] || \
-   [ "$GIT_USER_EMAIL" = "xtask@example.invalid" ]; then
-    echo "❌ Refusing commit with placeholder git identity"
-    echo "   user.name:  $GIT_USER_NAME"
-    echo "   user.email: $GIT_USER_EMAIL"
-    echo ""
-    echo "   Fix this repo-local override first:"
-    echo "   git config --local --unset-all user.name"
-    echo "   git config --local --unset-all user.email"
-    exit 1
-fi
-"#;
-    write_git_hook(&hooks_dir.join("pre-commit"), pre_commit_hook)?;
-
-    write_git_hook(&hooks_dir.join("pre-push"), pre_push_hook_script())?;
-
-    println!("✅ Installed pre-commit and pre-push hooks");
-    println!("   The pre-commit hook blocks known placeholder git identities");
-    println!("   The pre-push hook runs 'nix develop -c just pr-fast' before each push");
-    println!("   Skip with: git commit --no-verify / git push --no-verify");
-    Ok(0)
-}
-
-fn resolve_git_hooks_dir(repo_root: &Path) -> Result<PathBuf> {
-    let output = Command::new("git")
-        .current_dir(repo_root)
-        .args(["rev-parse", "--git-path", "hooks"])
-        .output()
-        .with_context(|| format!("resolving git hooks dir from {}", repo_root.display()))?;
-
-    if !output.status.success() {
-        return Err(color_eyre::eyre::eyre!(
-            "git rev-parse --git-path hooks failed in {}: {}",
-            repo_root.display(),
-            String::from_utf8_lossy(&output.stderr).trim_end()
-        ));
-    }
-
-    let hooks_path = String::from_utf8(output.stdout)
-        .context("git rev-parse --git-path hooks emitted non-UTF8 output")?;
-    let hooks_path = hooks_path.trim();
-    let hooks_dir = PathBuf::from(hooks_path);
-
-    Ok(if hooks_dir.is_absolute() { hooks_dir } else { repo_root.join(hooks_dir) })
-}
-
-fn write_git_hook(hook_path: &Path, hook: &str) -> Result<()> {
-    fs::write(hook_path, format!("{hook}\n"))
-        .with_context(|| format!("writing {:?}", hook_path))?;
-    #[cfg(unix)]
-    {
-        fs::set_permissions(hook_path, fs::Permissions::from_mode(0o755))
-            .with_context(|| format!("setting executable bit for {:?}", hook_path))?;
-    }
-    Ok(())
 }
 
 fn read_required_usize(path: &Path) -> Result<usize> {
